Rename quack_activity to quack_active_connections
