+++ src/CMakeLists.txt
@@ -2,6 +2,7 @@ add_subdirectory(storage)
 
 add_library(
   quack_library OBJECT
+  quack_active_connections.cpp
   quack_clear_cache.cpp
   quack_client.cpp
   quack_extension.cpp
@@ -13,7 +14,6 @@ add_library(
   quack_start_stop.cpp
   quack_storage.cpp
   quack_uri.cpp
-  quack_activity.cpp
   serialize_quack_message.cpp)
 set(EXTENSION_SOURCES
     ${EXTENSION_SOURCES} $<TARGET_OBJECTS:quack_library>
